General: plug HibernateWineRepository and fix some bugs.

+++ mycellar-java/mycellar-application/src/main/java/fr/peralta/mycellar/application/wine/CountryService.java
@@ -30,6 +30,6 @@ public interface CountryService {
     /**
      * @return
      */
-    Map<Country, Integer> getAllWithCounts();
+    Map<Country, Long> getAllWithCounts();
 
 }
